print.h: replace else..if with switch..case and use fput* functions
* Replace all else..if branches with switch..case. A optimizing
compiler might be able to create a lookup table instead relying
on the branches.

* Replace fprintf() with fputc() and fputs(), for strings that
do not require formatting. C compilers such as TCC (that doesn't
replace fprintf() calls, for (non-formatting) strings with fput*
functions) can make use of this. This shall increase the performance
a bit or so when compiled with TCC.